backend/requirements.txt
Django
djangorestframework
mssql-django
django-cors-headers
djangorestframework-simplejwt
django.contrib.sites 